+++ docs/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## UNRELEASED
+## v1.6.0
 
 ### New features